fix(xai): keep websocket compact helpers on the executor
Fold source selection and payload construction into
xai_websockets_executor.go so executor/ only contains executors and
tests. Do not put these helpers in helps/; they use executor-private
types and statusErr.
